requirements.txt
flask
tensorflow
numpy
opencv-python-headless
pillow
scikit-learn
joblib
xgboost
python-dotenv
google-genai
